Fix museum display case
