+++ data/area/misthalin/varrock/varrock_museum.ifaces.toml
@@ -5,7 +5,7 @@ id = 534
 id = 2
 
 [.vm_timeline_terracotta_statue_model]
-id = 49
+id = 50
 
 [.display_num]
 id = 84
