+++ game/src/main/kotlin/content/quest/member/the_golem/TheGolem.kt
@@ -9,8 +9,6 @@ import content.quest.questJournal
 import content.quest.questStage
 import world.gregs.voidps.engine.Script
 import world.gregs.voidps.engine.client.message
-import world.gregs.voidps.engine.client.sendScript
-import world.gregs.voidps.engine.client.ui.open
 import world.gregs.voidps.engine.data.definition.VariableDefinitions
 import world.gregs.voidps.engine.entity.character.move.tele
 import world.gregs.voidps.engine.entity.character.player.Player
@@ -51,16 +49,6 @@ class TheGolem : Script {
             sound("pick")
             anim("take")
             item("black_mushroom", "You pick a mushroom.")
-        }
-
-        // vm_timeline_no_display is the empty-case transform for every museum display,
-        // so scope both options to the terracotta statue's own base object.
-        objectOperate("Study", "vm_timeline_terracotta_statue_multi,vm_timeline_no_display") { (target) ->
-            if (target.id != "vm_timeline_terracotta_statue") {
-                noInterest()
-                return@objectOperate
-            }
-            studyDisplayCase()
         }
 
         objectOperate("Open", "vm_timeline_terracotta_statue_multi") {
@@ -165,27 +153,6 @@ class TheGolem : Script {
             inventory.add("golem_program")
             item("golem_program", "You write on the papyrus:<br>YOUR TASK IS DONE")
         }
-    }
-
-    private fun Player.studyDisplayCase() {
-        if (!open("vm_timeline")) {
-            return
-        }
-        anim("vm_display_case_ponder")
-        val stolen = get("golem_retrieved_statuette", false)
-        sendScript("museum_rotate_display", 0, 5, 0, 534 shl 16 or 50)
-        interfaces.sendModel("vm_timeline", "vm_timeline_terracotta_statue_model", if (stolen) 25568 else 25576)
-        interfaces.sendText("vm_timeline", "display_num", "30")
-        interfaces.sendText(
-            "vm_timeline",
-            "vm_timeline_text",
-            "3rd Age - yr 3000-4000<br><br>This " +
-                "statuette was found in an underground temple in the ruined city of Uzer, which was destroyed late " +
-                "in the 3rd Age, suddenly, due to causes unknown. It probably represents one of the clay golems " +
-                "that the craftsmen of the city built as warriors and servants. The statuette was originally part " +
-                "of a mechanism whose purpose is unknown." +
-                (if (stolen) "<br><br>Recently this display was stolen and its whereabouts are unknown." else ""),
-        )
     }
 
     private suspend fun Player.openDisplayCase() {
